arango-to-mongo-migration/requirements.txt
pymongo
dateutils
ijson